minor improvement
