+++ sql/hive/src/main/scala/org/apache/spark/sql/hive/HiveContext.scala
@@ -251,15 +251,11 @@ class HiveContext(sc: SparkContext) extends SQLContext(sc) {
   @transient
   override protected[sql] lazy val analyzer =
     new Analyzer(catalog, functionRegistry, caseSensitive = false) {
-      override val extendedResolutionRules =
+      override val extendedRules =
         catalog.CreateTables ::
         catalog.PreInsertionCasts ::
         ExtractPythonUdfs ::
         ResolveUdtfsAlias ::
-        Nil
-
-      override val extendedCheckRules =
-        CheckMultiAlias ::
         Nil
     }
